Moved outfits & locker below pool link
Moved the options to wear an outfit & the locker link below the 
"Enter the pool room" link, so it's easier to access the more 
important "enter" link.
Also swapped the "enter" and "leave" links when both the 
lockers & pool rooms are empty, to conform with how they are when 
either are occupied
Also moved & removed some breaks to be prettier.