Sydney Science preparations
Some code prep for implementing Sydney in Science class.
Also causes Pure Sydney to be more strict on the player skipping class to talk to them, and gives Sydney a cameo when the player interrupts Sirris' second period Science class.

This is my first real MR in a while, so sorry if there's any jank.