Remove obsolete warnings about history depth
Since the save size is no longer of concern, the warnings about history depth breaking quotas can be removed.
Additionally, a toggle is added to prevent saving the history into sessionStorage, so the performance penalty doesn't apply as well.
The toggle needs manual (de)activation to take any effect (for now), so it should be safe to include with the bugfixes.